4 commit
modify 2 servlets - ViewById and Delete - checking for ID

all servlets work

+++ src/main/java/com/example/demo/DeleteServlet.java
@@ -6,15 +6,32 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
 
 @WebServlet("/deleteServlet")
 public class DeleteServlet extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        String sid = request.getParameter("id");
-        int id = Integer.parseInt(sid);
-        EmployeeRepository.delete(id);
-        response.sendRedirect("viewServlet");
+        response.setContentType("text/html");
+        PrintWriter out = response.getWriter();
+
+        try {
+            String sid = request.getParameter("id");
+            int id = Integer.parseInt(sid);
+
+            Employee employee = EmployeeRepository.getEmployeeById(id);
+
+            if (employee.getId() == 0) {
+                throw new IOException("wrong ID");
+            }
+            EmployeeRepository.delete(id);
+//            out.print ("ID deleted successfully!");
+            response.sendRedirect("viewServlet");
+        }
+
+        catch (Throwable e) {
+                out.print ("not exist ID");
+            }
     }
 }
